requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
pandas==2.1.3
numpy==1.25.2
fake-useragent==1.4.0
python-dotenv==1.0.0
urllib3==2.0.7
certifi==2023.11.17 